requirements.txt
alembic==1.13.2
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
click==8.1.7
Flask==3.0.3
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
pillow==11.0.0
soupsieve==2.6
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
